+++ BE/src/stock/order/stock-order.service.ts
@@ -4,7 +4,6 @@ import {
   Injectable,
 } from '@nestjs/common';
 import { NotFoundError } from 'rxjs';
-import { Injectable } from '@nestjs/common';
 import { StockOrderRequestDto } from './dto/stock-order-request.dto';
 import { StockOrderRepository } from './stock-order.repository';
 import { TradeType } from './enum/trade-type';
